requirements.txt
langchain<0.3
langchain-community<0.3
langchain-groq
chromadb
streamlit
pandas
python-dotenv
sentence-transformers
langchain-huggingface